Change type of WindDirection to int (HourlyWeather)

+++ Models/HourlyWeather.cs
@@ -12,7 +12,7 @@ public class HourlyWeather
     public double Snowfall { get; set; } // mm
     public double Pressure { get; set; } // hPa
     public double WindSpeed { get; set; } // m/s
-    public double WindDirection { get; set; } // deg. (0-360)
+    public int WindDirection { get; set; } // deg. (0-360)
     public int Humidity { get; set; }
     public int WeatherCode { get; set; }
 }
